do not count the deployer

+++ barman/htdocs/act/launcher.cgi
@@ -94,11 +94,14 @@ class Launcher
       puts "Запускаю «#{job[1]}»…"
       fork { exec job[0] }
       Process.wait
-      error_file = Config::SAVE_ERROR % k
-      if $?.exitstatus == 0
-        File.unlink(error_file) if File.exists?(error_file)
-      else
-        File.write(error_file, @user_login)
+      
+      unless k == "deployer"
+        error_file = Config::SAVE_ERROR % k
+        if $?.exitstatus == 0
+          File.unlink(error_file) if File.exists?(error_file)
+        else
+          File.write(error_file, @user_login)
+        end
       end
     end
     unlock
